Use spacing vars, rename stuff as per review
