+++ src/components/Molecules/Footer/Nav/Nav.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 
 import Text from '../../../Atoms/Text/Text';
 import { sizes } from '../../../../theme/shared/breakpoint';
-import NavHelper from '../../../../utils/NavHelper';
+import NavHelper from '../../../../utils/navHelper';
 
 import {
   Nav,
