+++ src/components/Molecules/Header/Header.style.js
@@ -1,4 +1,4 @@
-// import React, { useState } from 'react';
+/* import React, { useState } from 'react';*/
 import styled from 'styled-components';
 
 import Link from '../../Atoms/Link/Link';
